move method that fetches GH data out of transformations
